refactored project
i modularized the app by moving the routes to separate files thus making the main file (now app.js) more readable.
