+++ routes/index.js
@@ -1,10 +1,8 @@
 const router = require('express').Router();
 
 // Visualize and manage the DB data
-router.get('/', function (req, res, next) {
-    res.render('admin', {
-
-    })
+router.get("/", function (req, res) {
+    res.render('index');
 });
 
 module.exports = router;
